help oneyeong

+++ beakJoon/greedyAlgorithm/beakjoon1202/beakjoon1202_usingHeap.c
@@ -107,7 +107,7 @@ int main(){
 			j++;
 		}
 		if(heapSize != 0){
-			answer += (long long int)getMaxFromHeap();
+			answer += getMaxFromHeap();
 		}
 	}
